+++ eq_cims_management_ui/main/routes.py
@@ -8,7 +8,7 @@ from flask import (
     request,
 )
 
-from eq_cims_management_ui.utils.database.firestore import FirestoreHandler
+from eq_cims_management_ui.utils.database.firestore_handler import FirestoreHandler
 
 main_blueprint = Blueprint("main", __name__)
 
@@ -41,8 +41,8 @@ def create_session() -> str:
         str: A rendered HTML page with a message indicating that the session was created. Note: To be updated to
         return a rendered page with a list of CIs.
     """
-    firestore_handler.create_session()
-    firestore_handler.read_latest_session()
+    created_session = firestore_handler.create_session()
+    firestore_handler.read_latest_session(created_session)
     return render_template("index.html", text="Session created")
